[iOS] Returning to fullscreen to PiP on YouTube breaks subsequent touch input https://bugs.webkit.org/show_bug.cgi?id=267674 rdar://119832557
Reviewed by NOBODY (OOPS!).

When returning to Element Fullscreen from PiP, the normal "exit picture-in-picture" logic
is interrupted in order to restore the previous element fullscreen. Recent changes have
left the video fullscreen state machine broken, waiting to restore the video to its original
location. The WKFullscreenViewController is left in place, but empty, either atop the WKWebView,
or below it but containing the video layer.

When we receive the preparedToReturnToStandby(), instead of just clearing the PiP fullscreen
mode, continue processing the steps of the "exit picture-in-picture" logic at the point at
which it was interrupted.

* Source/WebCore/platform/ios/VideoFullscreenInterfaceAVKit.mm:
(VideoFullscreenInterfaceAVKit::preparedToReturnToStandby):
(VideoFullscreenInterfaceAVKit::returnToStandby):